==> 2-results.Rmd <==
---
title: "VVG_product_results"
author: "Joe Hilgard"
date: "January 18, 2016"
output: 
  word_document:
    reference_docx: template_0.docx
---


<!--Note to self: "Democrat" is reference level for pol_orien_f-->


```{r setup, echo = F, include = F}
load("./Study_results.Rdata")
library(pscl)
library(psych)
library(magrittr)
library(dplyr)
library(BayesFactor)
library(ggplot2)
library(broom)

# function for printing p-values
printp = function(x) {
  if (x > 1) return("That ain't a p-value, buddy!")
  if (x == 1) return("*p* = 1.000")
  if (x < .001) return("*p* < .001")
  if (x >= .001 & x < 1) return(paste("*p* = ", substring(round(x, 3), 2), sep = ""))  
}
# function for printing b, t, p
fetch_stats = function(x, df) {
  paste("*b* = ", x[1], ", *t*(", df, ") = ", x[3], ", ", printp(x[4]), sep="")
}
# function for printing M and SD
m_sd = function(x) {
  paste0("*M* = ", round(mean(x, na.rm=T),2), 
        ", *SD* = ", round(sd(x, na.rm=T),2)
        )
}
# function for reporting alpha
alf = function(x) {
  x$total$raw_alpha %>% 
    round(2) %>% 
    substring(2)
}
```
``` {r cronbachs, echo = F, include = F}
# 1st amendment
alpha.1a <- dat %>% 
  select(speech_protect, game_amend, media_reg, ben_risk, glad_speech) %>% 
  psych::alpha()

# 2nd amendment
alpha.2a <- dat %>% 
  select(gun_protect, gun_amendment, gun_reg, glad_gun) %>% 
  psych::alpha()

# AR-15 desire
alpha.desire <- dat %>% 
  select(AR_1:AR_5) %>% 
  psych::alpha()

# Buying intent
alpha.intent <- dat %>% 
  select(AR_6:AR_8) %>% 
  psych::alpha()

# game-gun desire
cor.gamegun <- with(dat, cor(want, need))
ggplot(dat, aes(want, need)) +
  geom_jitter() + geom_smooth() 
# looks a bit of a dud -- most say no to both, esp need


# public policy
alpha.policy <- dat %>% 
  select(us_gun_laws:school_staff) %>% 
  psych::alpha(check.keys = T) # note reverse-scoring!
```
``` {r prepareCoefs, echo = F, include = F}
# Prepare finer points of models:
coef.cap = model.cap %>% 
  summary %$% 
  coefficients %>% 
  round(3)
df.cap = model.cap %$% df.residual

coef.cap_bin = model.cap_bin %>% 
  summary %$% 
  coefficients %>% 
  round(3)
df.cap_bin = model.cap_bin %$% df.residual

coef.deaths.count = model.deaths2 %>% 
  summary %$% 
  coefficients %$%
  count %>% 
  round(3)
coef.deaths.zero = model.deaths2 %>% 
  summary %$% 
  coefficients %$%
  zero %>% 
  round(3)
df.deaths = model.deaths %$% df.residual

coef.kills = model.kills2 %>% 
  summary %$% 
  coefficients %>%  
  round(3)
df.kills = model.kills2 %$% df.residual

coef.gamma1 = gamma1 %>% 
  summary %$% 
  coefficients %>% 
  round(3)
df.gamma1 = gamma1 %$% df.residual

coef.gamma2 = gamma2 %>% 
  summary %$% 
  coefficients %>% 
  round(3)
df.gamma2 = gamma2 %$% df.residual

coef.gamma3 = gamma3 %>% 
  summary %$% 
  coefficients %>% 
  round(3)
df.gamma3 = gamma3 %$% df.residual

coef.gamma4 = gamma4 %>% 
  summary %$% 
  coefficients %>% 
  round(3)
df.gamma4 = gamma4 %$% df.residual
```


**Manipulation check.** We tested how assignment to the 2 × 2 design influenced participants' in-game performance, as measured by the number of times the player died and the number of monsters the player killed. 

Count of player deaths was highly skewed and perhaps best described by a zero-inflated negative binomial distribution. A logit link was used for the zero inflation, and a log link was used for the negative binomial. Supporting this modeling decision, the overdispersion parameter was statistically significant, `r printp(summary(model.deaths2)$coefficients$count[5, 4])`. 
With regard to the number of times players died, neither the gun power (strong vs. weak; `r fetch_stats(coef.deaths.count["Power_fStrong",], df.deaths)`) nor gun type (realistic vs. sci-fi; `r fetch_stats(coef.deaths.count["Gun_type_fAR-15",], df.deaths)`) significantly influenced this outcome. Gun power and gun type did not significantly interact, `r fetch_stats(coef.deaths.count["Gun_type_fAR-15:Power_fStrong",], df.deaths)`.  
With regard to whether players died at all (e.g., the zero-inflation model parameters), neither the gun power (`r fetch_stats(coef.deaths.zero["Power_fStrong",], df.deaths)`) nor gun type (`r fetch_stats(coef.deaths.zero["Gun_type_fAR-15",], df.deaths)`) had a significant effect. They did not significantly interact, either (`r fetch_stats(coef.deaths.zero["Gun_type_fAR-15:Power_fStrong",], df.deaths)`).

Count of killed enemies was also highly skewed, and so a negative binomial distribution was fit using a log link. Participants in the powerful-gun condition killed substantially more enemies than did those in the weak-gun condition, `r fetch_stats(coef.kills["Power_fStrong",], df.kills)`, supporting the efficacy of the gun-power manipulation. Neither gun type (`r fetch_stats(coef.kills["Gun_type_fAR-15",], df.kills)`) nor the Gun Type × Gun Power interaction (`r fetch_stats(coef.kills["Gun_type_fAR-15:Power_fStrong",], df.kills)`) were significantly related to the number of enemies killed. 

Regrettably, we did not ask participants directly about how fun, powerful, satisfying, etc. the in-game gun was. As such, we do not have direct evidence that the powerful gun was more pleasant to use than the weak gun. At best, we might infer that, because the powerful gun was more effective, it was also more pleasant to use.  

Means and SDs of all variables are summarized in Table 2. Combination violin/boxplots for each outcome are summarized in Figures 2 and 3.

**2nd Amendment Advocacy.** Participants' 2nd Amendment advocacy was best modeled by a simple additive model of political orientation and gender, BF = `r exp(mod1_cov["pol_orien_f + Gender_f"]@bayesFactor$bf) %>% signif(3)` : 1 over the null. The covariates-only model was preferred over the full model (that is, the model with covariates, gun type, gun power, and the Gun Type × Gun Power interaction), BF = `r (exp(mod1_cov["pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod1_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` : 1. This indicates that the experimental condition had little explanatory power over and above that of the covariates. Comparison of the full model against the additive model let us examine the evidence for or against the hypothesized Gun Type × Gun Power interaction. The evidence was against this interaction, BF = 1 : `r (exp(mod1_cov["Gun_type_f + Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod1_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` for the full model relative to the additive-effects model.  

**Product attitudes.** Again, attitudes towards the AR-15 were best described by a simple additive model of political orientation and gender, BF = `r exp(mod5_cov["pol_orien_f + Gender_f"]@bayesFactor$bf) %>% signif(3)` : 1 over the null. The covariates-only model was preferred to the full model, BF = `r (exp(mod5_cov["pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod5_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` : 1. The evidence was against a Gun Type × Gun Power interaction, BF = 1 : `r (exp(mod5_cov["Gun_type_f + Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod5_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` for the full model relative to the additive-effects model.   

**Purchasing intentions.** Purchasing intentions were right-skewed. However, the QQplot of standardized residuals was not bad, and transforming intentions by square root or logarithm did not improve the QQplot much. Thus, we analyze this outcome in its natural units with the standard general linear modeling technique. 

Purchasing intentions were best described by additive effects of political orientation and gender, BF = `r exp(mod2_cov["pol_orien_f + Gender_f"]@bayesFactor$bf) %>% signif(3)` : 1 over the null. The covariates-only model was preferred to the full model, BF = `r (exp(mod2_cov["pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod2_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` : 1. The evidence was against a Gun Type × Gun Power interaction, BF = 1 : `r (exp(mod2_cov["Gun_type_f + Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod2_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` for the full model relative to the additive-effects model.    

**Desire of in-game weapon.** This variable was very badly right-skewed, with most participants choosing the minimum response. Square-root or log transformation did little to fix this. We report this in its natural units, but readers with better ideas for modeling are encouraged to use the raw data to perform further tests.

Weapon desire was best described by an effect of gender, BF = `r exp(mod3_cov["Gender_f"]@bayesFactor$bf) %>% signif(3)` : 1 over the null. Adding the main and interactive effects of gun type and gun power to this gender-only model was not preferred, BF = 1 : `r (exp(mod3_cov["Gender_f"]@bayesFactor$bf)/exp(mod3_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)`. The evidence was against the predicted Gun Type × Gun Power interaction, BF = 1 : `r (exp(mod3_cov["Gun_type_f + Power_f + Gender_f"]@bayesFactor$bf)/exp(mod3_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` for the full model relative to the additive-effects model.     

**Policy opinion.** Policy views were best described by political orientation and gender alone, BF = `r exp(mod4_cov["pol_orien_f + Gender_f"]@bayesFactor$bf) %>% signif(3)` : 1 over the null. Adding the main and interactive effects of gun type and gun power was not supported, BF = 1 : `r (exp(mod4_cov["pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod4_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)`. The evidence was against the hypothesized Gun Type × Gun Power interaction, BF = 1 : `r (exp(mod4_cov["Gun_type_f + Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)/exp(mod4_cov["Gun_type_f + Power_f + Gun_type_f:Power_f + pol_orien_f + Gender_f"]@bayesFactor$bf)) %>% signif(3)` for the full model relative to the additive-effects model.      

**Rates of gun accidents and gun use.** Participants' estimated rates seemed to be more appropriately modeled as a gamma distribution than a normal distribution. Because responses of 0% cannot be modeled under this distribution, these responses were adjusted to 0.001%. In general, we note that participants' estimates were highly variable, ranging from 0% to 80% or more.

Only a few idiosyncratic predictors reached statistical significance. Republicans, relative to liberals, thought it more probable that a gun owner would experience a gun-related accident such as an accidental discharge, `r fetch_stats(coef.gamma1["pol_orien_fRep",], df.gamma1)`. Men, relative to women, thought it more probable that a gun owner might have a gun stolen from them, `r fetch_stats(coef.gamma2["Gender_fMale",], df.gamma2)`. Libertarians, relative to other political parties, thought it more probable that a gun owner would ever use their gun in an act of self-defense, `r fetch_stats(coef.gamma4["pol_orien_fLib",], df.gamma3)`. None of these estimated rates were significantly predicted by the game participants had played. Full tables of model output are provided in Supplementary Table 1.

**Magazine capacity.** Several participants listed very large values (e.g., 100 or more) for a maximum magazine size, or wrote in responses to the effect that there should be no such government-imposed limit. We tried modeling this outcome in two ways. First, we winsorized all responses in excess of 30 down to 30 and attempted a linear model. Second, we coded a dichotomous variable for responses less than 30 and responses equal to or greater than 30 and attempted a logistic model. The linear model suggested that males (`r fetch_stats(coef.cap["Gender_fMale",], df.cap)`) and libertarians (`r fetch_stats(coef.cap["pol_orien_fLib",], df.cap)`) supported larger magazine sizes than did females and democrats. The logistic model found no significant predictors. Neither model detected any significant effects of game.
